+++ sardine/handlers/superdirt.py
@@ -7,7 +7,13 @@ from osc4py3.as_eventloop import osc_send, osc_udp_client
 
 from ..utils import alias_param
 from .osc_loop import OSCLoop
-from .sender import Number, NumericElement, ParsableElement, Sender, StringElement
+from .sender import (
+    Number,
+    NumericElement,
+    ParsableElement,
+    Sender,
+    StringElement
+)
 
 __all__ = ("SuperDirtHandler",)
 
@@ -123,34 +129,6 @@ class SuperDirtHandler(Sender):
         }
         return rename_keys(pattern, aliases)
 
-    def cycle_should_play(
-            self,
-            struct: Optional[tuple | int],
-            on: Optional[tuple | int],
-    ):
-        """Determine if a function should play based
-        on a cycle count from origin."""
-
-        # Always return True if the mechanism is not in use
-        if struct is None or on is None:
-            return True
-
-        # If we only give an int, build the struct from 0
-        if isinstance(struct, int):
-            struct = (1, struct)
-
-        bar = self.env.clock.bar
-        group = struct[1]
-
-        elapsed_bars = bar // group
-        bar_in_current_group = bar - (elapsed_bars * group)
-
-        if isinstance(on, tuple):
-            return True if bar_in_current_group in tuple(x - 1 for x in on) else False
-        elif isinstance(on, int):
-            return bar_in_current_group == on
-
-
 
     @alias_param(name="iterator", alias="i")
     @alias_param(name="divisor", alias="d")
@@ -168,7 +146,9 @@ class SuperDirtHandler(Sender):
             return
     
         # If the result of this cycle computation is false, we don't have to play at all
-        if not self.cycle_should_play(pattern.get("struct", None), pattern.get("on", None)):
+        if not self.cycle_should_play(
+                pattern.get("struct", None),
+                pattern.get("on", None)):
             return
 
         # Replace some shortcut parameters by their real name
@@ -205,6 +185,12 @@ class SuperDirtHandler(Sender):
     ) -> int | float:
         # Replace some shortcut parameters by their real name
         pattern = self._parse_aliases(pattern)
+
+        # If the result of this cycle computation is false, we don't have to play at all
+        if not self.cycle_should_play(
+                pattern.get("struct", None),
+                pattern.get("on", None)):
+            return
 
         if not self._ziffers_parser:
             raise Exception("The ziffers package is not imported!")
